Buys and sells stocks

+++ Core/Entities/IEXModels/QuoteModel.cs
@@ -13,9 +13,9 @@ namespace Core.Entities.IEXModels
 
         public decimal previousClose { get; set; }
 
-        public decimal change { get; set; }
+        public double change { get; set; }
 
-        public decimal changePercent { get; set; }
+        public double changePercent { get; set; }
 
         public decimal week52High { get; set; }
